requirements.txt
django
django-menu
django-tastypie